Adding sample about middlewares after routes
