+++ 06-express-api/xpress-api-03/index.js
@@ -27,22 +27,18 @@ router.use(function(req, res, next) {
 });
 
 // test route to make sure everything is working (accessed at GET http://localhost:8080/api)
-router.get('/', function(req, res) {
-    res.json({ message: 'hooray! welcome to our api!' });   
+router.get('/', function(req, res, next) {
+    res.json({ message: 'hooray! welcome to our api! Look how middleware is executed!' });  
+    next(); 
 });
 
-// more routes for our API will happen here
-// ----------------------------------------------------
-router.route('/persons') //please lets use plural
+router.use( function( req, res, next ) {
 
-    // create a bear (accessed at POST http://localhost:8080/api/bears)
-    .post(function(req, res) {
-        console.log('Person ready');
-        var enrique = new Person();      // create a new instance of the Person model
-        enrique.name = req.body.name;  // set the person name (comes from the request)
-        
-        res.json({ message: 'Person created!' });
-    });
+    console.log( "after" );
+    next();
+  
+  } );
+// more routes for our API will happen here
 
 // REGISTER OUR ROUTES -------------------------------
 // all of our routes will be prefixed with /api
@@ -52,10 +48,3 @@ app.use('/api', router);
 // =============================================================================
 app.listen(port);
 console.log('Magic happens on port ' + port);
-
-
-class Person{
-    eat(){
-        console.log('eat...');
-    }
-}
